Base price calculation added. Markup generated moved to markup service. Recalculate base price updated. Article base price cache updated.

+++ src/Core/Core/Models/MarkupModel.cs
@@ -1,3 +1,3 @@
 ﻿namespace Core.Models;
 
-public record MarkupModel(double Markup);
+public record MarkupModel(decimal Markup);
